# Add zlib-ng for more efficient compression at higher levels