--- a/setup.cfg
+++ b/setup.cfg
@@ -19,6 +19,7 @@ package_dir =
 packages = find:
 install_requires =
     isal>=1.0.0; platform.python_implementation == 'CPython' and (platform.machine == "x86_64" or platform.machine == "AMD64")
+    zlib-ng>=0.1.0; platform.machine == "x86_64" or platform.machine == "AMD64" or platform.machine == "aarch64"
     typing_extensions; python_version<'3.8'
 
 [options.packages.find]