fixes #119