@rhpvorderman 2023-02-01
Wrote a little benchmarking script (saving it here for reuse later):
```Python3
#!/usr/bin/env python3
import itertools
import os
import sys
import time
from xopen import xopen


def main():
    testfile = sys.argv[1]
    out = sys.argv[2]
    for threads, compresslevel in itertools.product((0, 1, 4), range(1, 7)):
        start = time.time()
        block_size = 128 * 1024
        with open(testfile, "rb") as in_file:
            with xopen(
                out, "wb", compresslevel=compresslevel, threads=threads
            ) as out_file:
                while True:
                    block = in_file.read(block_size)
                    if block == b"":
                        break
                    out_file.write(block)

        stop = time.time()
        print(
            f"Compressed at level {compresslevel} with {threads} threads; "
            f"filesize: {os.stat(out).st_size}, "
            f"time: {stop -start} seconds"
        )


if __name__ == "__main__":
    main()
```
Current main:
```
Compressed at level 1 with 0 threads; filesize: 405313510, time: 3.3767712116241455 seconds
Compressed at level 2 with 0 threads; filesize: 404989250, time: 3.6654465198516846 seconds
Compressed at level 3 with 0 threads; filesize: 407253483, time: 7.312707424163818 seconds
Compressed at level 4 with 0 threads; filesize: 386930500, time: 25.895344734191895 seconds
Compressed at level 5 with 0 threads; filesize: 374207207, time: 46.744099140167236 seconds
Compressed at level 6 with 0 threads; filesize: 355347468, time: 109.18307518959045 seconds
Compressed at level 1 with 1 threads; filesize: 405313274, time: 3.204677104949951 seconds
Compressed at level 2 with 1 threads; filesize: 404989061, time: 3.268627405166626 seconds
Compressed at level 3 with 1 threads; filesize: 407251397, time: 7.039590835571289 seconds
Compressed at level 4 with 1 threads; filesize: 387016236, time: 25.5415096282959 seconds
Compressed at level 5 with 1 threads; filesize: 374290891, time: 46.197908878326416 seconds
Compressed at level 6 with 1 threads; filesize: 355500797, time: 108.11184501647949 seconds
Compressed at level 1 with 4 threads; filesize: 405313274, time: 3.209529399871826 seconds
Compressed at level 2 with 4 threads; filesize: 404989061, time: 3.3253917694091797 seconds
Compressed at level 3 with 4 threads; filesize: 407251397, time: 7.039556980133057 seconds
Compressed at level 4 with 4 threads; filesize: 387016236, time: 6.912431478500366 seconds
Compressed at level 5 with 4 threads; filesize: 374290891, time: 12.286827325820923 seconds
Compressed at level 6 with 4 threads; filesize: 355500797, time: 28.226975440979004 seconds
```
Results zlib-ng branch:
```
Compressed at level 1 with 0 threads; filesize: 405313510, time: 3.390442132949829 seconds
Compressed at level 2 with 0 threads; filesize: 404989250, time: 3.4523391723632812 seconds
Compressed at level 3 with 0 threads; filesize: 407253483, time: 7.442785978317261 seconds
Compressed at level 4 with 0 threads; filesize: 368802018, time: 21.94339346885681 seconds
Compressed at level 5 with 0 threads; filesize: 359170597, time: 28.51734709739685 seconds
Compressed at level 6 with 0 threads; filesize: 347379910, time: 47.41038107872009 seconds
Compressed at level 1 with 1 threads; filesize: 405313274, time: 3.212735652923584 seconds
Compressed at level 2 with 1 threads; filesize: 404989061, time: 3.277775287628174 seconds
Compressed at level 3 with 1 threads; filesize: 407251397, time: 7.07685923576355 seconds
Compressed at level 4 with 1 threads; filesize: 368802018, time: 20.918381452560425 seconds
Compressed at level 5 with 1 threads; filesize: 359170597, time: 27.791999101638794 seconds
Compressed at level 6 with 1 threads; filesize: 347379910, time: 44.957818031311035 seconds
Compressed at level 1 with 4 threads; filesize: 405313274, time: 3.197093963623047 seconds
Compressed at level 2 with 4 threads; filesize: 404989061, time: 3.2572927474975586 seconds
Compressed at level 3 with 4 threads; filesize: 407251397, time: 7.0963521003723145 seconds
Compressed at level 4 with 4 threads; filesize: 368802018, time: 20.83365035057068 seconds
Compressed at level 5 with 4 threads; filesize: 359170597, time: 27.362520933151245 seconds
Compressed at level 6 with 4 threads; filesize: 347379910, time: 44.91855430603027 seconds
```

Some observations: 
+ zlib-ng obtains nice results at single-threaded and 0-threaded use cases especially regarding filesize.
+ not having multithreaded options for zlib-ng hurts a bit. Crabz should come next as suggested by #117 .  
+ Igzip at level 3 is a bit redundant. It only works effectively when AVX-512 is available. This is not available usually, so I'd rather have zlib-ng run this level. This however is not possible with the current model, where we simply try if it works and use the first option that sticks.
+ Pigz at 4 threads pretty much crushes zlib-ng at 1 thread (as expected) so it might be better to prefer that when more threads are an option and crabz is not available? (Though unlikely, as it is available in conda and I also think that is the only reason people have pigz available.)

I think when adding crabz the code will be even more crowded. So a different model should be used in order to properly choose the correct tool for the job weighing threads, compression level etc. But that is for another PR.