@rhpvorderman 2023-02-02
This leaves a few holes in the test suite as what willl happen when zlib-ng is not installed is not properly tested. I will address these in a coming PR where I will refactor _open_gz a bit as it is currently getting a bit too overcomplicated. 
If you prefer, I can also do it as part of this PR.